% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMIDNet_20230328V2.R
\name{calMIDSim}
\alias{calMIDSim}
\title{calMIDSim}
\usage{
calMIDSim(
  NodeTable = NULL,
  Pattern_name = "Unknown",
  TargetTable_dir = "./QuantTableContiminateAdjustScaled.csv",
  TargetSample = NULL,
  dir_path = ".",
  Cutoff = 0.9,
  IsoDegree = 0.1,
  FoldChange = 20,
  MinMotifLength = 0.5,
  multi_thread = 1,
  maxMotif = 10000,
  ignore_max = FALSE
)
}
\arguments{
\item{TargetTable_dir}{Stable isotope labeling quantification table. CSV table. First 1-6 column: name, id, mz, compound, label. The label column: 0-n represent isotopologue M0 - Mn}

\item{Cutoff}{MID similarity score. Default: 0.9}

\item{IsoDegree}{minimum labeling extent. Note: the labeled isotopologue can't be reduced in similarity scoring. Default: 0.1}

\item{FoldChange}{a parameter to judge similarity scoring type. If the most abundant isotopologue divide the second abundant isotopologue larger than this value, the isotopologue simialrity is type I, else the type is type II. Default: 20}

\item{MinMotifLength}{Character. Minimum length of motif (value < 1 means fraction of max motif length). Default: 0.5}

\item{maxMotif}{the maximum motif combination. Default: 10000}
}
\description{
calMIDSim
}
\author{
Hongmiao Wang, Zhiwei Zhou
}
