% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunMs2Net.R
\name{calMS2Net}
\alias{calMS2Net}
\title{calMS2Net}
\usage{
calMS2Net(
  network_file = "GlobalMatchResult_filterred.csv",
  ms2_file = "ms2_data.msp",
  dir_path = ".",
  scoring_approach = "gnps",
  mz_tol_ms2 = 15,
  ms2_score_cutoff = 0.5,
  num_matched_frag_cutoff = 1
)
}
\description{
calMS2Net
}
\author{
Zhiwei Zhou
}
