% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{constructMIDNet}
\alias{constructMIDNet}
\title{constructMIDNet}
\usage{
constructMIDNet(
  tracer_table_file = "MID_more_than_3_POS_maxC.csv",
  ms2_file = "ms2_data.msp",
  id_table = "id_table.csv",
  sample_names,
  dir_path = ".",
  multi_thread = 1,
  mid_cutoff = 0.95,
  mid_fc = 20,
  mid_isoDegree = 0.02,
  mid_min_motifLen = 0.6,
  mid_max_motif = 1000,
  ignore_max = FALSE,
  scoring_approach = "gnps",
  mz_tol_ms2 = 15,
  ms2_score_cutoff = 0.5,
  num_matched_frag_cutoff = 1,
  mz_tol_label = 15,
  mass_diff_freq_cut_off = 1
)
}
\description{
constructMIDNet
}
\author{
Zhiwei Zhou
}
